% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDDPhase.R
\name{dFuncGierullEq7}
\alias{dFuncGierullEq7}
\title{Probability density function}
\usage{
dFuncGierullEq7(x, param)
}
\arguments{
\item{x}{Equivalent to \code{sample=psi}, is the array describing the support of the
phase at which the pdf  is computed.}

\item{r}{Denotes the magnitude of the complex correlation (coherence),
\code{0<r<=1}.}

\item{theta}{Is a potential mean between, \code{-pi<theta<=pi}.}

\item{L}{Number of looks or targeted, \code{L>1} integer.}
}
\value{
A graphic
}
\description{
This function calculates the pdf of the interferometric phase.
             Article "Closed-form expressions for InSAR sample statistics and
             its application to non-gaussian data" Gierull (2020), Eq. 13.
}
\details{
The correct phase pdf is derived using identity tables for GHF.
}
\examples{
sample = seq(-pi,pi, 0.001)
param <- c(r = 0.9, theta = 0, L = 4)
data<-data.frame(x=sample,y=dFuncGierullEq7(sample, param))
ggplot(data,aes(x,y)) +
  geom_line() +
  xlim(-pi, pi) +
  # coord_polar() +
  xlab("sample") +
  ylab("density") +
  theme_pander()
}
\author{
Araújo, J. E.
}
