% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDDPhase.R
\name{dFuncLeeEq5}
\alias{dFuncLeeEq5}
\title{Probability density function}
\usage{
dFuncLeeEq5(x, param)
}
\arguments{
\item{x}{Equivalent to \code{sample=psi}, is the array describing the support of the
phase at which the pdf  is computed.}

\item{r}{Complex correction coefficient (coherence), \code{0<r<=1}.}

\item{theta}{Is a potential mean between, \code{-pi<theta<=pi}.}

\item{L}{Number of looks or target, \code{L>=1} positive real.}
}
\value{
A graphic
}
\description{
Article "Intensity and phase statistics of multilook polarimetric
             and interferometric SAR imagery", Lee et al. (1994), Eq. 18.
}
\details{
Multilook probability density function for phase difference
         involving the Hypergeometric Gauss Function (GHF).
}
\examples{
sample <- seq(-pi,pi, 0.001)
param <- c(r = 0.7, theta = 0, L = 4)
data<-data.frame(x=sample,y=dFuncLeeEq5(sample, param))
ggplot(data,aes(x,y)) +
  geom_line() +
  xlim(-pi, pi) +
  # coord_polar() +
  xlab("sample") +
  ylab("density") +
  theme_pander()
}
\author{
Araújoo, J. E.
}
