% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDDPhase.R
\name{dTruncNorm}
\alias{dTruncNorm}
\title{Probability density function}
\usage{
dTruncNorm(x, param)
}
\arguments{
\item{psi}{Equivalent to \code{sample=psi}, is the array describing the support of the
phase at which the pdf.}

\item{mu}{loc: Floating point tensor; the means of the distribution(s).}

\item{sigma}{scale: loating point tensor; the stddevs of the distribution(s).
Must contain only positive values.}
}
\value{
A graphic
}
\description{
The truncated normal is a normal distribution bounded between
             low and high (the pdf is zero outside these bounds and
             renormalized). Samples from this distribution are differentiable
             with respect to loc, scale as well as the bounds,low and high,
             i.e., this implementation is fully reparameterizeable
}
\details{
In probability and statistics, the truncated normal distribution is
         the probability distribution derived from that of a normally
         distributed random variable by bounding the random variable from
         either below or above (or both). The truncated normal distribution
         has wide applications in statistics and econometrics.
}
\examples{
sample = seq(-pi,pi, 0.001)
param <- c(mu = 0, sigma = 0.37)
data<-data.frame(x=sample,y=dTruncNorm(sample, param))
ggplot(data,aes(x,y)) +
geom_line() +
xlab("sample") +
ylab("density") +
scale_x_continuous (breaks = c(-pi, -pi/2, 0, pi/2, pi),
labels = c(expression(-pi), expression(-pi/2), expression(0),
expression(pi/2), expression(pi)))+ theme(legend.position = "none",
axis.title.x = element_text(size = 12),
axis.title.y = element_text(size = 12))
}
\author{
Araújo, J. E.
}
