% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDDPhase.R
\name{estim.GierullEq7}
\alias{estim.GierullEq7}
\title{Maximization with the maxLik package}
\usage{
estim.GierullEq7(sample, param)
}
\description{
If it is necessary to maximize the parameters of the dFuncGierullEq7,
             density, we indicate the maxLik package.
}
\details{
The main function of the maxLik package is to perform the Maximum
         Likelihood estimate.

maxLik::maxLik(LogLikelihoodGierullEq7, start=c(r, theta, L))
}
\examples{
## Estimate the parameters of Gierull (2020), Eq. 13. Density.
sample <- c( 0.21850138, 0.74192327, 1.42014964, 0.03348661, 0.65550808,
0.46397016, 0.03027485)
param <- c(r = 0.7, theta = 0, L = 5)
mle.Gierull <- estim.GierullEq7(sample, param)
summary(mle.Gierull)

}
