% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDDPhase.R
\name{estim.LeeEq5}
\alias{estim.LeeEq5}
\title{Maximization with the maxLik package}
\usage{
estim.LeeEq5(sample, param)
}
\description{
If it is necessary to maximize the parameters of the dFuncLeeEq4,
             density, we indicate the maxLik package.

maxLik::maxLik(LogLikelihoodLeeEq5, start=c(r, theta,L))
}
\details{
The main function of the maxLik package is to perform the Maximum
         Likelihood estimate.
}
\examples{
## Estimate the parameters of Lee et al. (1994), Eq. 18 Density.
sample <- c( 0.21850138, 0.74192327, 1.42014964, 0.03348661, 0.65550808,
0.46397016, 0.03027485)
param <- c(r = 0.7, theta = 0, L = 5)
mle.Lee <- estim.LeeEq5(sample, param)
summary(mle.Lee)
}
