% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDDPhase.R
\name{estim.TruncCauchy}
\alias{estim.TruncCauchy}
\title{Maximization with the maxLik package}
\usage{
estim.TruncCauchy(sample, param)
}
\description{
If it is necessary to maximize the parameters of the dTruncCauchy,
             density, we indicate the maxLik package.
}
\details{
The main function of the maxLik package is to perform the Maximum
         Likelihood estimate.

maxLik::maxLik(LogLikelihoodTCauchy, start=c(mu, sigma))
}
\examples{
## Estimate the parameters of Truncated Cauchy Density
sample <- c( 0.21850138, 0.74192327, 1.42014964, 0.03348661, 0.65550808,
0.46397016, 0.03027485)
param <- c(mu = 0.5, sigma = 1)
mle.TruncCauchy <- estim.TruncCauchy(sample, param)
summary(mle.TruncCauchy)

}
