% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDDPhase.R
\name{dTruncCauchy}
\alias{dTruncCauchy}
\title{Probability density function}
\usage{
dTruncCauchy(x, param)
}
\arguments{
\item{x}{Equivalent to \code{sample=psi}, is the array describing the support of the
phase at which the pdf.}

\item{mu}{loc: Floating point tensor; the modes of the corresponding
non-truncated Cauchy distribution(s).}

\item{sigma}{scale: Floating point tensor; the scales of the distribution(s).
Must contain only positive values.}
}
\value{
A graphic
}
\description{
The truncated Cauchy is a Cauchy distribution bounded between
             low a nd high (the pdf is 0 outside these bounds and renormalized).
             Samples from this distribution are differentiable with respect to
             loc and scale, but not with respect to the bounds low and high
}
\details{
The truncated Cauchy distribution has the same density, with a
         different proportionality constant, over a finite interval. It's a
         well-behaved distribution, with a mean, variance, and any other
         moment you might care about.
}
\examples{
sample = seq(-pi,pi, 0.001)
param <- c(mu = 0, sigma = 0.21)
data<-data.frame(x=sample,y=dTruncCauchy(sample, param))
ggplot(data,aes(x,y)) +
geom_line() +
xlab("sample") +
ylab("density") +
scale_x_continuous (breaks = c(-pi, -pi/2, 0, pi/2, pi),
labels = c(expression(-pi), expression(-pi/2), expression(0),
expression(pi/2), expression(pi)))+ theme(legend.position = "none",
axis.title.x = element_text(size = 12),
axis.title.y = element_text(size = 12))
}
\author{
Araújo, J. E.
}
